refactor: create views folder for organizing components better
