+++ pages/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import HomePage from '../components/HomePage';
+import HomePage from '../views/HomePage';
 
 export default function Home() {
   return <HomePage />
