+++ pages/pokedex/tracker.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import BasicTracker from '../../components/Pokedex/BasicTrackerPage';
+import BasicTracker from '../../views/Pokedex/BasicTrackerPage';
 
 export default function Tracker() {
   return <BasicTracker />
