+++ views/FIRTrackerPage/FIRTrackerPage.jsx
@@ -2,7 +2,7 @@ import classnames from 'classnames/bind';
 import Head from 'next/head';
 import React from 'react';
 import Layout from '../../components/Layout';
-import UnderConstruction from '../UnderConstruction';
+import UnderConstruction from '../../components/UnderConstruction';
 import styles from './FIRTrackerPage.module.css';
 
 const cx = classnames.bind(styles);
