+++ views/Pokedex/MinesweeperPage/index.js
